+++ test/csp.js
@@ -7,15 +7,17 @@ var a = require("../src/csp.test-helpers"),
 
 var csp = require("../src/csp"),
     chan = csp.chan,
-    promiseChan = csp.promiseChan,
     go = csp.go,
     put = csp.put,
     take = csp.take,
     putAsync = csp.putAsync,
     takeAsync = csp.takeAsync,
+    offer = csp.offer,
+    poll = csp.poll,
     alts = csp.alts,
     timeout = csp.timeout,
-    CLOSED = csp.CLOSED;
+    CLOSED = csp.CLOSED,
+    NO_VALUE = csp.NO_VALUE;
 
 var do_alts = require("../src/impl/select").do_alts;
 
@@ -43,13 +45,6 @@ describe("put", function() {
     it("should return false if channel is already closed", function*() {
       var ch = closed(chan);
       assert.equal((yield put(ch, 42)), false);
-    });
-
-    it("should always return true when putting to a promise channel", function*() {
-      var ch = promiseChan();
-      assert.equal((yield put(ch, 42)), true);
-      assert.equal((yield put(ch, 43)), true);
-      assert.equal((yield put(ch, 44)), true);
     });
   });
 
@@ -82,15 +77,6 @@ describe("put", function() {
 
     //   assert.equal(buffered, true, "pending put is buffered once the buffer is not full again");
     // });
-
-    it("should return true if value is then taken from promise channel", function*() {
-      var ch = promiseChan();
-      go(function*() {
-        yield timeout(5);
-        yield take(ch);
-      });
-      assert.equal((yield put(ch, 42)), true);
-    });
 
     it("should return false if channel is then closed", function*() {
       var ch = chan();
@@ -154,19 +140,8 @@ describe("take", function() {
       assert.equal((yield take(ch)), 42);
     });
 
-    it("should return correct value that was first put in promise channel", function*() {
-      var ch = promiseChan();
-      yield put(ch, 42);
-      assert.equal((yield take(ch)), 42);
-    });
-
     it("should return false if channel is already closed", function*() {
       var ch = closed(chan);
-      assert.equal((yield take(ch)), CLOSED);
-    });
-
-    it("should return false if promise channel is already closed", function*() {
-      var ch = closed(promiseChan);
       assert.equal((yield take(ch)), CLOSED);
     });
   });
@@ -181,18 +156,6 @@ describe("take", function() {
       assert.equal((yield take(ch)), 42);
     });
 
-    it("should return correct value if it is then delivered to promise channel", function*() {
-      var ch = promiseChan();
-      go(function*() {
-        yield timeout(5);
-        yield put(ch, 42);
-        yield put(ch, 43);
-        yield put(ch, 44);
-      });
-      assert.equal((yield take(ch)), 42);
-      assert.equal((yield take(ch)), 42);
-    });
-
     it("should return CLOSED if channel is then closed", function*() {
       var ch = chan();
 
@@ -203,17 +166,53 @@ describe("take", function() {
 
       assert.equal((yield take(ch)), CLOSED);
     });
+  });
+});
 
-    it("should return CLOSED if promise channel is then closed", function*() {
-      var ch = promiseChan();
+describe("offer and poll", function() {
+  function noOp() {};
 
-      go(function*() {
-        yield timeout(5);
-        ch.close();
-      });
+  mocha.it("should succeed if they can be completed immediately by a buffer", function() {
+    var ch = chan(2);
+    assert.equal(offer(ch, 42), true);
+    assert.equal(offer(ch, 43), true);
+    assert.equal(offer(ch, 44), false);
+    assert.equal(poll(ch), 42);
+    assert.equal(poll(ch), 43);
+    assert.equal(poll(ch), NO_VALUE);
+  });
 
-      assert.equal((yield take(ch)), CLOSED);
-    });
+  mocha.it("should succeed if they can be completed immediately by a pending operation", function() {
+    var putCh = chan();
+    putAsync(putCh, 42);
+    assert.equal(poll(putCh), 42);
+
+    var takeCh = chan();
+    takeAsync(takeCh, noOp);
+    assert.equal(offer(takeCh, 42), true);
+  });
+
+  mocha.it("should fail if they can't complete immediately", function() {
+    var ch = chan();
+    assert.equal(poll(ch), NO_VALUE);
+    assert.equal(offer(ch, 44), false);
+  });
+
+  mocha.it("should fail if they are performed on a closed channel", function() {
+    var ch = chan();
+    ch.close();
+    assert.equal(poll(ch), NO_VALUE);
+    assert.equal(offer(ch, 44), false);
+  });
+
+  mocha.it("should fail if there are pending same-direction operations on a channel", function() {
+    var putCh = chan();
+    putAsync(putCh, 42);
+    assert.equal(offer(putCh, 44), false);
+
+    var takeCh = chan();
+    takeAsync(takeCh, noOp);
+    assert.equal(poll(takeCh), NO_VALUE);
   });
 });
 
@@ -499,30 +498,6 @@ describe("close", function() {
 
   it("should correctly flush CLOSED to pending takes", function*() {
     var ch = chan();
-    var count = 0;
-
-    go(function*() {
-      assert.equal((yield take(ch)), CLOSED);
-      count += 1;
-      assert.equal(count, 1);
-    });
-    go(function*() {
-      assert.equal((yield take(ch)), CLOSED);
-      count += 1;
-      assert.equal(count, 2);
-    });
-    go(function*() {
-      assert.equal((yield take(ch)), CLOSED);
-      count += 1;
-      assert.equal(count, 3);
-    });
-
-    ch.close();
-    yield undefined;
-  });
-
-  it("should correctly flush CLOSED to pending takes from promise channel", function*() {
-    var ch = promiseChan();
     var count = 0;
 
     go(function*() {